Refactor for testing
